separation of instructionStatements function
