+++ Urho3D/splash_build/CMakeFiles/Splash.dir/depend.internal
@@ -8,6 +8,12 @@ CMakeFiles/Splash.dir/fft.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
  /home/pi/sound-pirates/Urho3D/splash/fft.cpp
  /home/pi/sound-pirates/Urho3D/splash/lol.h
+CMakeFiles/Splash.dir/instructionsStatements.cpp.o
+ /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/fft.h
+ /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.cpp
+ /home/pi/sound-pirates/Urho3D/splash/lol.h
+ /home/pi/sound-pirates/Urho3D/splash/playNote.h
 CMakeFiles/Splash.dir/playNote.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
  /home/pi/sound-pirates/Urho3D/splash/fft.h
@@ -170,6 +176,7 @@ CMakeFiles/Splash.dir/splash.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/Sample.h
  /home/pi/sound-pirates/Urho3D/splash/Sample.inl
  /home/pi/sound-pirates/Urho3D/splash/fft.h
+ /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
  /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.h
  /home/pi/sound-pirates/Urho3D/splash/splash.cpp
