+++ Urho3D/splash_build/CMakeFiles/Splash.dir/progress.make
@@ -4,4 +4,5 @@ CMAKE_PROGRESS_3 = 3
 CMAKE_PROGRESS_4 = 4
 CMAKE_PROGRESS_5 = 5
 CMAKE_PROGRESS_6 = 6
+CMAKE_PROGRESS_7 = 7
